Consolidate all code into one file
