+++ main_page/src/add_definition.py
@@ -9,8 +9,8 @@ def main():
     result = "drmf_bof\n" + \
              "'''Definition:" + macro_name + "'''\n" + \
              "The LaTeX DLMF and DRMF macro '''\\" + macro_name + "''' represents the " + macro_name + ".\n\n" + \
-             "These are defined by \n\n" + \
-             "<math>\n\n{\\displaystyle\n\n" + definition + "\n\n}</math>\n\n<br />\n\ndrmf_eof\n"
+             "These are defined by \n" + \
+             "<math>{\\displaystyle\n" + definition + "\n}</math>\n<br />\n\ndrmf_eof\n"
 
     with open("main_page.mmd", "a") as main_page:
         main_page.write("\n" + result)
